pyproject.toml
[project]
name = "durak"
version = "0.1.0"
description = ""
authors = [
    {name = "@Semka_96",email = "max.skuybida@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "id (>=1.6.1,<2.0.0)",
    "nanoid (>=2.0.0,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "syrupy (>=5.1.0,<6.0.0)",
]
package-mode = false

[tool.black]
include = '(\.py$)'
line-length = 88
target-version = ['py313']
skip-string-normalization = true
preview = false


[tool.poetry.group.dev.dependencies]
black = "^26.3.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
